feat: AlarmController 생성
issue: #234
