+++ app/src/main/java/com/nbcfinalteam2/ddaraogae/presentation/ui/alarm/AlarmController.kt
@@ -1,0 +1,17 @@
+package com.nbcfinalteam2.ddaraogae.presentation.ui.alarm
+
+import android.app.AlarmManager
+
+class AlarmController(
+    private val alarmManager: AlarmManager
+) {
+
+    fun setAlarm(id: Int, setTime: Int) {
+        
+    }
+
+    fun deleteAlarm(id: Int) {
+
+    }
+
+}
