+++ app/src/main/java/com/nbcfinalteam2/ddaraogae/presentation/ui/alarm/AlarmSetDialogFragment.kt
@@ -2,7 +2,6 @@ package com.nbcfinalteam2.ddaraogae.presentation.ui.alarm
 
 import android.app.AlarmManager
 import android.content.Context.ALARM_SERVICE
-import android.content.Intent
 import android.graphics.Color
 import android.graphics.drawable.ColorDrawable
 import android.os.Bundle
@@ -12,9 +11,7 @@ import android.view.ViewGroup
 import android.view.Window
 import androidx.fragment.app.DialogFragment
 import com.nbcfinalteam2.ddaraogae.databinding.FragmentAlarmSetDialogBinding
-import com.nbcfinalteam2.ddaraogae.presentation.receiver.AlarmReceiver
 import dagger.hilt.android.AndroidEntryPoint
-import java.util.Calendar
 
 @AndroidEntryPoint
 class AlarmSetDialogFragment(
@@ -42,19 +39,9 @@ class AlarmSetDialogFragment(
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
         binding.btnConfirm.setOnClickListener {
-            val hour = binding.tpAlarm.hour
-            val minute = binding.tpAlarm.minute
             alarmDialogButtonListener.onPositiveButtonClicked(
                 binding.tpAlarm.hour*60 + binding.tpAlarm.minute
             )
-
-            val intent = Intent(requireContext().applicationContext, AlarmReceiver::class.java)
-            intent.putExtra("set_time", binding.tpAlarm.hour*60 + binding.tpAlarm.minute)
-            val calendar = Calendar.getInstance().apply {
-                timeInMillis = System.currentTimeMillis()
-                set(Calendar.HOUR_OF_DAY, hour)
-                set(Calendar.MINUTE, minute)
-            }
             dismiss()
         }
         binding.btnCancel.setOnClickListener {
